feat(feature/ujin/#56): 친구 타입, 유저 정보 타입 types 폴더에 분리
