+++ yeoreum/src/api/friendPage.ts
@@ -8,13 +8,6 @@ export const config = {
   },
 };
 
-export type FriendResponseType = {
-  friendUserNo: number;
-  friendNickname: string;
-  friendDescription: string;
-  friendProfileImage: string;
-}[];
-
 export const fetchSearchFriends = async (value: string) => {
   const { data } = await remote.get(`/api/friends/${value}`, config);
   return data.response;
