+++ yeoreum/src/types/user.ts
@@ -1,0 +1,10 @@
+export interface UserProfileResponseType {
+  userNo: number;
+  email: string;
+  nickname: string;
+  major: string;
+  gender: number;
+  description: string;
+  profileImage: string;
+  grade: string;
+}
